fix some numeric types cast issues
Former-commit-id: 64475cac441527ee994d0d9aeb5993e070a0ba4c [formerly c199d9fdf2d806cc4d9d92725e6bcdcc51bd2334] [formerly 721aa3935dc8338494c60c00d7a4e48514f6f635]
Former-commit-id: 04861d9e18375897281c34e52f2414b4b7b6aade [formerly 2f709f53af9d93697d4280906bf0a04fa3700bf0] [formerly c436a313a5ad32f678664acb681c81be94d3bed3]
Former-commit-id: d7e3dbdaccf1b6d1d0262f6d50e397c8c0d8e508
Former-commit-id: bdffde9f7c9984a9084bdeba22fb9e24e4a37b05

+++ plugins/bullet/src/minko/component/bullet/SphereShape.hpp
@@ -67,7 +67,7 @@ namespace minko
 				volume() const
 				{
 					const float radius	= _radius + _margin;
-					const float volume	= (PI * 4.0f / 3.0f) * radius * radius * radius;
+					const float volume	= ((float)PI * 4.0f / 3.0f) * radius * radius * radius;
 					
 					return volume * _volumeScaling;
 				}
